requirements.txt
appdirs==1.4.4
feedparser==6.0.1
lxml==4.5.2
opml==0.5
python-dateutil==2.8.1
sgmllib3k==1.0.0
six==1.15.0
youtube-dl==2020.9.20
